+++ src/test/java/edu/illinois/library/cantaloupe/resource/InformationRequestHandlerTest.java
@@ -61,7 +61,7 @@ public class InformationRequestHandlerTest extends BaseTest {
             final Configuration config = Configuration.getInstance();
             config.setProperty(Key.SOURCE_STATIC, "FilesystemSource");
             config.setProperty(Key.FILESYSTEMSOURCE_PATH_PREFIX,
-                    TestUtil.getImagesPath().toString() + "/");
+                    TestUtil.getImagesPath() + "/");
         }
 
         final IntrospectiveCallback callback = new IntrospectiveCallback();
@@ -80,7 +80,7 @@ public class InformationRequestHandlerTest extends BaseTest {
             final Configuration config = Configuration.getInstance();
             config.setProperty(Key.SOURCE_STATIC, "FilesystemSource");
             config.setProperty(Key.FILESYSTEMSOURCE_PATH_PREFIX,
-                    TestUtil.getImagesPath().toString() + "/");
+                    TestUtil.getImagesPath() + "/");
         }
 
         final IntrospectiveCallback callback = new IntrospectiveCallback();
@@ -100,7 +100,7 @@ public class InformationRequestHandlerTest extends BaseTest {
             config.setProperty(Key.CACHE_SERVER_RESOLVE_FIRST, false);
             config.setProperty(Key.SOURCE_STATIC, "FilesystemSource");
             config.setProperty(Key.FILESYSTEMSOURCE_PATH_PREFIX,
-                    TestUtil.getImagesPath().toString() + "/");
+                    TestUtil.getImagesPath() + "/");
             config.setProperty(Key.DERIVATIVE_CACHE_ENABLED, true);
             config.setProperty(Key.DERIVATIVE_CACHE, "HeapCache");
         }
@@ -131,12 +131,55 @@ public class InformationRequestHandlerTest extends BaseTest {
     }
 
     @Test
+    void testHandleSetsRequestContextPageCountBeforeReturningInstanceFromDerivativeCache()
+            throws Exception {
+        {   // Configure the application.
+            final Configuration config = Configuration.getInstance();
+            config.setProperty(Key.CACHE_SERVER_RESOLVE_FIRST, false);
+            config.setProperty(Key.SOURCE_STATIC, "FilesystemSource");
+            config.setProperty(Key.FILESYSTEMSOURCE_PATH_PREFIX,
+                    TestUtil.getImagesPath() + "/");
+            config.setProperty(Key.DERIVATIVE_CACHE_ENABLED, true);
+            config.setProperty(Key.DERIVATIVE_CACHE, "HeapCache");
+            config.setProperty(Key.DELEGATE_SCRIPT_ENABLED, true);
+            config.setProperty(Key.DELEGATE_SCRIPT_PATHNAME,
+                    TestUtil.getFixture("delegates.rb").toString());
+        }
+
+        // Configure the request.
+        final Identifier identifier = new Identifier("jpg-rgb-64x48x8.jpg");
+        final Metadata metadata     = new Metadata();
+
+        // Add an info to the derivative cache.
+        CacheFacade facade = new CacheFacade();
+        DerivativeCache cache = facade.getDerivativeCache().orElseThrow();
+        Info info = Info.builder()
+                .withSize(64, 48)
+                .withFormat(Format.get("jpg"))
+                .withIdentifier(identifier)
+                .withMetadata(metadata)
+                .build();
+        cache.put(identifier, info);
+
+        final IntrospectiveCallback callback = new IntrospectiveCallback();
+        final RequestContext context = new RequestContext();
+        try (InformationRequestHandler handler = InformationRequestHandler.builder()
+                .withCallback(callback)
+                .withIdentifier(identifier)
+                .withRequestContext(context)
+                .build()) {
+            handler.handle();
+            assertEquals(1, context.getPageCount());
+        }
+    }
+
+    @Test
     void testHandleReturnsInstanceFromProcessor() throws Exception {
         {   // Configure the application.
             final Configuration config = Configuration.getInstance();
             config.setProperty(Key.SOURCE_STATIC, "FilesystemSource");
             config.setProperty(Key.FILESYSTEMSOURCE_PATH_PREFIX,
-                    TestUtil.getImagesPath().toString() + "/");
+                    TestUtil.getImagesPath() + "/");
         }
 
         final IntrospectiveCallback callback = new IntrospectiveCallback();
@@ -146,6 +189,28 @@ public class InformationRequestHandlerTest extends BaseTest {
                 .build()) {
             Info info = handler.handle();
             assertNotNull(info);
+        }
+    }
+
+    @Test
+    void testHandleSetsRequestContextPageCountBeforeReturningInstanceFromProcessor()
+            throws Exception {
+        {   // Configure the application.
+            final Configuration config = Configuration.getInstance();
+            config.setProperty(Key.SOURCE_STATIC, "FilesystemSource");
+            config.setProperty(Key.FILESYSTEMSOURCE_PATH_PREFIX,
+                    TestUtil.getImagesPath() + "/");
+        }
+
+        final IntrospectiveCallback callback = new IntrospectiveCallback();
+        final RequestContext context = new RequestContext();
+        try (InformationRequestHandler handler = InformationRequestHandler.builder()
+                .withCallback(callback)
+                .withIdentifier(new Identifier("jpg-rgb-64x48x8.jpg"))
+                .withRequestContext(context)
+                .build()) {
+            handler.handle();
+            assertEquals(1, context.getPageCount());
         }
     }